django_react/pyproject.toml
[tool.poetry]
name = "django-react"
version = "0.1.0"
description = ""
authors = ["M NOMAN CH <noman.chaudhary@tylextech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
flower = "^2.0.1"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
